Add `PartialMonoid` typeclass
It is sometimes useful to consider partial semigroups that also have
an identity element, say `pmempty`.  Many, but not all, partial
semigroups that have instances in this library are also partial
monoids, with some identity element.  This patch introduces a
typeclass, `PartialMonoid`, that represents such types.  This
typeclass is meant to directly mirror the standard `Monoid` typeclass,
prepending `pm` to each function in the standard typeclass and
changing their types to represent partiality.  The only deviation here
is removing `pmappend` from the typeclass, since this is necessarily
an alias for `<>?` specialized to monoids; unlike with the standard
`mappend`, by moving this out of the typeclass, we avoid any
possibility of unlawful instances where `pmappend ≢ (<>?)`.

This patch tags each addition with a `@since 0.7.0.0` Haddock
annotation, on the assumption that these changes will be included in a
new release of the library with that version.